pci: refactor: use smallest types where possible
When working with PCI, types used to represent different parts
of the spec can use small integer types like u8/u16 instead of
usize. Usage of smaller types makes it harder to make mistakes
in the code and prevents accidental overflows.
This commit tries to replace most of the PCI used types to the
smallest ones which are enough to represent the valid range of
values required. This prevents the `try_from().unwrap()` logic
which can be error prone. Instead it only tries to widen the
type if necessary with safe `::from` functions. In minority of
cases the `#[allow(clippy::cast_possible_truncation)]` is used
with an additional `assert` or other check is used to ensure the
value is within correct range.

No functional changes intended.

+++ src/vmm/src/device_manager/pci_mngr.rs
@@ -120,7 +120,7 @@ impl PciDevices {
             .pci_bus
             .lock()
             .expect("Poisoned lock")
-            .add_device(bdf.device() as u32, virtio_device.clone());
+            .add_device(bdf.device(), virtio_device.clone());
 
         self.virtio_devices
             .insert((device_type, id), virtio_device.clone());
